requirements.txt
google-genai
pillow
pyautogui
pytesseract
opencv-python
numpy
mss
keyboard
pytest
pytest-mock
python-dotenv
black
flake8
bandit
isort
requests
pyinstaller
detect-secrets
flask
flask-cors
cryptography
